+++ src/App.tsx
@@ -1,9 +1,11 @@
 import React from "react";
 import "./App.css";
-import LoginPage from "../src/components/login/LoginPage";
+import LoginPage from "../src/screens/login/LoginPage";
 
 function App() {
-  return <LoginPage />;
+  return (
+      <LoginPage />
+  );
 }
 
 export default App;
